requirements.txt
discord.py==2.3.2
requests==2.31.0
aiohttp==3.9.4.post1
